rename cli to pack
